streamlit/requirements.txt
streamlit>=1.45.0
google-cloud-bigquery>=3.40.1
pandas>=2.0.0
